Update to handle dynamic authentication methods
The Facebook module is purely a placeholder right now to demonstrate
how it would be setup. It is disabled within the default config.

Test script for the auth module code:

  from pyrpgwnn import app
  print(app.config['AUTHS'])

This should result in a dict with a single key 'local' - to create
an object that handles authentication for the 'local' method, use:

  am = app.config['AUTHS']['local']['class']()

From the am object, you can call the methods defined within the
local Auth class which should eventually allow functionality such
as changing passwords, logging in, etc. Some functionality for
these auth methods is likely to be provided within the relevant
view module, for example the 'local' module would include routes for

  /register/local
  /login/local
  /login/local/pwrecover

whereas 'facebook' would not include the pwrecover route as that
is provided by Facebook's own website.